Multiple calls to wiringXISR work without loosing files
If you call wiringXISR to configure a pin for interrupt mode
and want to change the configuration again the sysFds[i]
file descriptor is overwritten without being closed.
You can force the close by calling wiringXGC. But I think it would
be nicer if nothing gets lost. And you don't need to close the
file if you just want to reconfigure the pin.