New approach to modular SoC / Platform support
The new wiringX is built around SoCs instead of
platforms. As in the real world, a platform is
merely a mask around a SoC. So in this rewrite
a start was made in fully mapping each SoC used
for each platform supported. This makes adding
new platforms easy, because we just have to map
the memory addresses of the SoCs to the physical
GPIO that the platforms expose.

The new wiringX intentionally maps all GPIO
addresses, instead of using all kinds of formulas.
This is done for easier porting and mapping by
less experiences developers.

Functions such as SPI, I2C and Serial are all
implemented in wiringX itself instead of for
each platform individually. Auto platform
detection was also removed because it proved to
be too unreliable.

This new based was built around two yet
unsupported platforms: the PCDuino1 with the
Allwinner A10 and the BananaPi M2 with the
Allwinner A31s. The Allwinner SoCs were choosen
as a first prove of concept because a lot of
platform use it.